Create a generalized JITLinker
